+++ src/shared/entities/Comment.interface.ts
@@ -1,4 +1,4 @@
-import { Publishable } from './Publishable.js';
+import { Publishable } from './utils/Publishable.interface.js';
 
 export interface Comment extends Publishable {
   text: string;
