+++ src/shared/entities/Place.interface.ts
