+++ src/shared/entities/User.interface.ts
